+++ swift-paperless/Views/Document/Detail/V4/DocumentDetailViewV4.swift
@@ -11,353 +11,6 @@ import Flow
 import Networking
 import SwiftUI
 
-extension Common.SchemeToken where Value == Color {
-  fileprivate static var editButtonColor: Self {
-    .init(light: Color(white: 0.3), dark: Color(white: 0.5))
-  }
-
-  fileprivate static var backgroundColor: Self {
-    .init(light: Color(white: 0.97), dark: Color(white: 0.2))
-  }
-}
-
-private enum TransitionID: Hashable {
-  case doc
-  case tags
-  case date
-  case asn
-  case correspondent
-  case documentType
-  case storagePath
-  case owner
-  case metadata
-  case notes
-}
-
-private struct EditableAspect: View {
-  let label: String?
-  let systemImage: String
-  let transitionID: TransitionID?
-  let namespace: Namespace.ID?
-  let action: (() -> Void)?
-  let showPrivateFallback: Bool
-  let accessibilityLabel: String?
-
-  @ScaledMetric(relativeTo: .body)
-  private var fontSizeRaw = 15
-
-  private var fontSize: CGFloat {
-    min(fontSizeRaw, 20)
-  }
-
-  private var pillPadding: CGFloat {
-    1 + fontSize / 6
-  }
-
-  private var iconPadding: CGFloat {
-    2 + fontSize / 6
-  }
-
-  @SchemeValue(.backgroundColor)
-  private var backgroundColor
-
-  @SchemeValue(light: Color(white: 0.9), dark: Color(white: 0.35))
-  private var iconBackgroundColor: Color
-
-  @SchemeValue(.editButtonColor)
-  private var editButtonColor
-
-  init(
-    localized: LocalizedStringResource, systemImage: String, action: (() -> Void)? = nil,
-    transitionID: TransitionID? = nil,
-    namespace: Namespace.ID? = nil,
-    showPrivateFallback: Bool = true,
-    accessibilityLabel: String? = nil
-  ) {
-    self.label = String(localized: localized)
-    self.systemImage = systemImage
-    self.action = action
-    self.transitionID = transitionID
-    self.namespace = namespace
-    self.showPrivateFallback = showPrivateFallback
-    self.accessibilityLabel = accessibilityLabel
-  }
-
-  init(
-    _ label: String?, systemImage: String, action: (() -> Void)? = nil,
-    transitionID: TransitionID? = nil,
-    namespace: Namespace.ID? = nil,
-    showPrivateFallback: Bool = true,
-    accessibilityLabel: String? = nil
-  ) {
-    self.label = label
-    self.systemImage = systemImage
-    self.action = action
-    self.transitionID = transitionID
-    self.namespace = namespace
-    self.showPrivateFallback = showPrivateFallback
-    self.accessibilityLabel = accessibilityLabel
-  }
-
-  private var displayLabel: String? {
-    if let label {
-      return label
-    }
-    if showPrivateFallback {
-      return String(localized: .permissions(.private))
-    }
-    return nil
-  }
-
-  private var computedAccessibilityLabel: String? {
-    accessibilityLabel ?? displayLabel
-  }
-
-  var body: some View {
-    Button {
-      action?()
-    } label: {
-      HStack {
-        Image(systemName: systemImage)
-          .resizable()
-          .scaledToFit()
-          .frame(width: fontSize, height: fontSize)
-          .padding(iconPadding)
-          .background(Circle().fill(iconBackgroundColor))
-          .padding(.vertical, pillPadding)
-          .padding(.leading, pillPadding)
-        if let displayLabel {
-          Text(displayLabel)
-            .italic(label == nil && showPrivateFallback)
-        }
-        Image(systemName: "pencil")
-          .foregroundStyle(editButtonColor)
-          .padding(.trailing, 2 + fontSize / 2)
-      }
-      .font(.system(size: fontSize))
-      .background {
-        Capsule()
-          .fill(backgroundColor)
-      }
-      .dynamicTypeSize(...DynamicTypeSize.large)
-      .apply {
-        if let transitionID, let namespace {
-          $0.backport.matchedTransitionSource(id: transitionID, in: namespace)
-        } else {
-          $0
-        }
-      }
-    }
-    .accessibilityLabel(computedAccessibilityLabel ?? "")
-    .buttonStyle(.plain)
-  }
-}
-
-private struct DocumentTagsSection: View {
-  @SchemeValue(.editButtonColor)
-  private var editButtonColor
-
-  @SchemeValue(.backgroundColor)
-  private var backgroundColor
-
-  @SchemeValue(light: Color(white: 0.9), dark: Color(white: 0.3))
-  private var editButtonBackground
-
-  let tags: [Tag?]
-  let action: (() -> Void)?
-  let transitionID: TransitionID?
-  let namespace: Namespace.ID?
-
-  var body: some View {
-    HStack(alignment: .top) {
-      Button {
-        action?()
-      } label: {
-        TagsView(
-          tags: tags, action: nil,
-          content: {
-            Label(.localizable(.edit), systemImage: "pencil")
-              .foregroundStyle(editButtonColor)
-              .apply {
-                if tags.isEmpty {
-                  $0.labelStyle(.titleAndIcon)
-                    .padding(EdgeInsets(top: 2, leading: 8, bottom: 2, trailing: 8))
-                } else {
-                  $0.labelStyle(.iconOnly)
-                    .padding(5)
-                }
-              }
-              .background {
-                if tags.isEmpty {
-                  Capsule()
-                    .fill(backgroundColor)
-                } else {
-                  Circle()
-                    .fill(backgroundColor)
-                }
-              }
-          }
-        )
-      }
-      .buttonStyle(.plain)
-      .frame(maxWidth: .infinity, alignment: .leading)
-    }
-    .dynamicTypeSize(...DynamicTypeSize.large)
-    .apply {
-      if let transitionID, let namespace {
-        $0.backport.matchedTransitionSource(id: transitionID, in: namespace)
-      } else {
-        $0
-      }
-    }
-  }
-}
-
-private struct CorrespondentEditSheet: View {
-  @Bindable var viewModel: DocumentDetailModel
-
-  @EnvironmentObject private var store: DocumentStore
-  @EnvironmentObject private var errorController: ErrorController
-  @Environment(\.dismiss) private var dismiss
-
-  @State private var searchText = ""
-  @State private var saving = false
-  @State private var selectedDetent: PresentationDetent = .medium
-
-  private var correspondents: [Correspondent] {
-    let search = searchText.lowercased()
-    return store.correspondents.values
-      .filter { search.isEmpty || $0.name.lowercased().contains(search) }
-      .sorted { $0.name.localizedCaseInsensitiveCompare($1.name) == .orderedAscending }
-  }
-
-  private var selectedCorrespondent: Correspondent? {
-    viewModel.document.correspondent.flatMap { store.correspondents[$0] }
-  }
-
-  private func select(correspondent: UInt?) {
-    guard !saving else { return }
-    guard viewModel.document.correspondent != correspondent else {
-      dismiss()
-      return
-    }
-
-    Task {
-      do {
-        saving = true
-        var document = viewModel.document
-        document.correspondent = correspondent
-        async let updatedDocument = store.updateDocument(document)
-        async let delay: () = Task.sleep(for: .seconds(0.3))
-
-        let updated = try await updatedDocument
-        try await delay
-        viewModel.document = updated
-        saving = false
-        dismiss()
-      } catch {
-        saving = false
-        errorController.push(error: error)
-      }
-    }
-  }
-
-  private func row(_ label: String, id: UInt?) -> some View {
-    Button {
-      select(correspondent: id)
-    } label: {
-      HStack {
-        Text(label)
-          .foregroundStyle(.primary)
-        Spacer()
-        if viewModel.document.correspondent == id {
-          Label(String(localized: .localizable(.elementIsSelected)), systemImage: "checkmark")
-            .labelStyle(.iconOnly)
-        }
-      }
-      .contentShape(Rectangle())
-    }
-    .buttonStyle(.plain)
-    .disabled(saving)
-  }
-
-  var body: some View {
-    NavigationStack {
-      ScrollView(.vertical) {
-        VStack(spacing: 0) {
-          CustomSection {
-            if let selectedCorrespondent {
-              Button {
-                select(correspondent: nil)
-              } label: {
-                HStack {
-                  Text(selectedCorrespondent.name)
-                    .foregroundStyle(.primary)
-                  Spacer()
-                  Image(systemName: "xmark.circle.fill")
-                    .foregroundStyle(.secondary)
-                }
-                .contentShape(Rectangle())
-              }
-              .buttonStyle(.plain)
-              .disabled(saving)
-            } else {
-              Text(.localizable(.correspondentNotAssignedPicker))
-                .foregroundStyle(.secondary)
-            }
-          } header: {
-            Text(.localizable(.selected))
-          }
-
-          CustomSection {
-            VStack(spacing: 0) {
-              ForEach(Array(correspondents.enumerated()), id: \.element.id) {
-                index, correspondent in
-                row(correspondent.name, id: correspondent.id)
-
-                if index < correspondents.count - 1 {
-                  Divider()
-                }
-              }
-            }
-          } header: {
-            Text(.localizable(.correspondents))
-          }
-        }
-      }
-      .customSectionBackgroundStyle(selectedDetent == .large ? .solid : .translucent)
-      .scrollBounceBehavior(.basedOnSize)
-      .searchable(text: $searchText, placement: .navigationBarDrawer(displayMode: .always))
-      .navigationTitle(.localizable(.correspondent))
-      .navigationBarTitleDisplayMode(.inline)
-      .toolbar {
-        ToolbarItem(placement: .cancellationAction) {
-          CancelIconButton()
-        }
-        ToolbarItem(placement: .confirmationAction) {
-          if saving {
-            ProgressView()
-          }
-        }
-      }
-    }
-    .presentationDetents([.medium, .large], selection: $selectedDetent)
-    .interactiveDismissDisabled(saving)
-    .apply {
-      if #available(iOS 26.0, *) {
-        if selectedDetent == .large {
-           $0.presentationBackground(Color(.systemGroupedBackground))
-//          $0.presentationBackground(.red)
-        } else {
-          $0
-        }
-      } else {
-        $0
-      }
-    }
-  }
-}
-
 @MainActor
 struct DocumentDetailViewV4: DocumentDetailViewProtocol {
   @State private var viewModel: DocumentDetailModel
@@ -805,199 +458,6 @@ struct DocumentDetailViewV4: DocumentDetailViewProtocol {
   }
 }
 
-private struct TagsEditSheet: View {
-  @Bindable var viewModel: DocumentDetailModel
-
-  @EnvironmentObject private var store: DocumentStore
-  @EnvironmentObject private var errorController: ErrorController
-
-  @Environment(\.dismiss) private var dismiss
-
-  @State private var tagIds: [UInt] = []
-  @State private var searchText = ""
-  @State private var saving = false
-  @State private var selectedDetent: PresentationDetent = .medium
-  @State private var showCreateTag = false
-
-  @Namespace private var tagNamespace
-
-  private struct CreateTagView: View {
-    @EnvironmentObject private var store: DocumentStore
-    @EnvironmentObject private var errorController: ErrorController
-    @Environment(\.dismiss) private var dismiss
-
-    let onCreated: (Tag) -> Void
-
-    var body: some View {
-      TagEditView<ProtoTag>(onSave: { value in
-        Task {
-          do {
-            let tag = try await store.create(tag: value)
-            onCreated(tag)
-            dismiss()
-          } catch {
-            errorController.push(error: error)
-          }
-        }
-      })
-      .navigationTitle(Text(.localizable(.tagCreateTitle)))
-      .navigationBarTitleDisplayMode(.inline)
-    }
-  }
-
-  private var interactiveDismissDisabled: Bool {
-    showCreateTag || tagIds != viewModel.document.tags
-  }
-
-  private var availableTags: [Tag] {
-    let search = searchText.lowercased()
-    return store.tags.values
-      .filter { !tagIds.contains($0.id) }
-      .filter { search.isEmpty || $0.name.lowercased().contains(search) }
-      .sorted { $0.name.localizedCaseInsensitiveCompare($1.name) == .orderedAscending }
-  }
-
-  private func save() {
-    Task {
-      do {
-        saving = true
-        var document = viewModel.document
-        document.tags = tagIds
-        let updated = try await store.updateDocument(document)
-        viewModel.document = updated
-        saving = false
-        dismiss()
-      } catch {
-        saving = false
-        errorController.push(error: error)
-      }
-    }
-  }
-
-  let animation = Animation.spring(duration: 0.2)
-
-  var body: some View {
-    NavigationStack {
-      ScrollView(.vertical) {
-        CustomSection {
-          if tagIds.isEmpty {
-            Text(.localizable(.noTagsSelected))
-              .foregroundStyle(.secondary)
-              .padding(.vertical, 8)
-              .frame(maxWidth: .infinity, alignment: .leading)
-          } else {
-            HFlow {
-              ForEach(tagIds, id: \.self) { tagId in
-                Button {
-                  withAnimation(animation) {
-                    tagIds.removeAll { $0 == tagId }
-                  }
-                } label: {
-                  TagView(tag: store.tags[tagId]) {
-                    Image(systemName: "xmark")
-                      .font(.caption2)
-                      .fontWeight(.bold)
-                  }
-                  .fixedSize()
-                }
-                .buttonStyle(.plain)
-                .matchedGeometryEffect(id: tagId, in: tagNamespace)
-              }
-            }
-            .padding(.vertical, 8)
-            .frame(maxWidth: .infinity, alignment: .topLeading)
-          }
-        }
-
-        if !availableTags.isEmpty {
-          CustomSection {
-            VStack(spacing: 0) {
-              ForEach(Array(availableTags.enumerated()), id: \.element.id) { index, tag in
-                Button {
-                  withAnimation(animation) {
-                    tagIds.append(tag.id)
-                  }
-                } label: {
-                  HStack {
-                    TagView(tag: tag)
-                      .fixedSize()
-                      .matchedGeometryEffect(id: tag.id, in: tagNamespace)
-                    Spacer()
-                    Image(systemName: "plus")
-                      .foregroundStyle(.secondary)
-                  }
-                  .padding(.vertical, 8)
-                  .contentShape(Rectangle())
-                }
-                .buttonStyle(.plain)
-
-                if index < availableTags.count - 1 {
-                  Divider()
-                }
-              }
-            }
-          }
-        }
-      }
-      .customSectionBackgroundStyle(selectedDetent == .large ? .solid : .translucent)
-      .scrollBounceBehavior(.basedOnSize)
-      .searchable(text: $searchText, placement: .navigationBarDrawer(displayMode: .always))
-      .navigationTitle(.localizable(.tags))
-      .navigationBarTitleDisplayMode(.inline)
-      .toolbar {
-        ToolbarItem(placement: .cancellationAction) {
-          CancelIconButton()
-        }
-        ToolbarItem(placement: .topBarTrailing) {
-          Button {
-            showCreateTag = true
-          } label: {
-            Label(String(localized: .localizable(.tagAdd)), systemImage: "plus")
-          }
-          .disabled(!store.permissions.test(.add, for: .tag))
-        }
-        ToolbarItem(placement: .confirmationAction) {
-          if saving {
-            ProgressView()
-          } else {
-            SaveButton {
-              save()
-            }
-            .fontWeight(.bold)
-            .disabled(tagIds == viewModel.document.tags)
-          }
-        }
-      }
-      .navigationDestination(isPresented: $showCreateTag) {
-        CreateTagView(onCreated: { tag in
-          withAnimation(animation) {
-            tagIds.append(tag.id)
-          }
-        })
-        .onDisappear {
-          showCreateTag = false
-        }
-      }
-    }
-    .presentationDetents([.medium, .large], selection: $selectedDetent)
-    .interactiveDismissDisabled(interactiveDismissDisabled)
-    .apply {
-      if #available(iOS 26.0, *) {
-        if selectedDetent == .large {
-          $0.presentationBackground(Color(.systemGroupedBackground))
-        } else {
-          $0
-        }
-      } else {
-        $0
-      }
-    }
-    .onAppear {
-      tagIds = viewModel.document.tags
-    }
-  }
-}
-
 // MARK: - Previews
 
 extension Tag {
@@ -1077,6 +537,9 @@ private struct DocumentDetailViewV4PreviewHelper: View {
 
         let correspondent = try await repository.create(
           correspondent: ProtoCorrespondent(name: "Some bank")
+        )
+        _ = try await repository.create(
+          correspondent: ProtoCorrespondent(name: "Health Insurance Co.")
         )
         let documentType = try await repository.create(
           documentType: ProtoDocumentType(name: "Preview Type")
